# tp2/src/openapi.yaml
openapi: 3.1.0
info:
  title: API de depósito
  version: "1.0.0"
  description: |
    Categorías y productos, lugares (depósito, pasillo, fila, estante, bin)
    y stock solo en bins. Un ingreso asigna la cantidad escaneada a uno o más
    bins y la suma una sola vez al completarse. Un pedido se pickea desde bins
    hacia un cartón y se embala cuando cada línea está completa.
servers:
  - url: http://localhost:3000
...
paths:
...
  /products:
    get:
      tags: [Productos]
      operationId: listProducts
      summary: Listar productos o buscar por SKU
      parameters:
        - name: sku
          in: query
          required: false
          description: Escaneo. Coincidencia exacta del SKU.
          schema:
            type: string
            minLength: 1
      responses:
        "200":
          description: Productos. Si se pasa sku y no hay coincidencia, la lista va vacía.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
        "400":
          $ref: "#/components/responses/BadRequest"
    post:
      tags: [Productos]
      operationId: createProduct
      summary: Crear un producto
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductWrite"
      responses:
        "201":
          description: Producto creado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /products/{productId}:
    parameters:
      - $ref: "#/components/parameters/productId"
    get:
      tags: [Productos]
      operationId: getProduct
      summary: Leer un producto
      responses:
        "200":
          description: Producto
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [Productos]
      operationId: updateProduct
      summary: Cambiar SKU, nombre o categoría
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductPatch"
      responses:
        "200":
          description: Producto actualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
    delete:
      tags: [Productos]
      operationId: deleteProduct
      summary: Borrar un producto sin stock, ingresos ni pedidos
      responses:
        "204":
          description: Borrado
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /products/{productId}/stock:
    parameters:
      - $ref: "#/components/parameters/productId"
    get:
      tags: [Productos]
      operationId: getProductStock
      summary: Saldos del producto por bin
      responses:
        "200":
          description: Saldos. La cantidad no se edita por acá.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductStock"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /orders:
    get:
      tags: [Pedidos]
      operationId: listOrders
      summary: Listar pedidos
      responses:
        "200":
          description: Pedidos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      tags: [Pedidos]
      operationId: createOrder
      summary: Abrir un pedido
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderWrite"
      responses:
        "201":
          description: Pedido abierto
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
  /orders/{orderId}:
    parameters:
      - $ref: "#/components/parameters/orderId"
    get:
      tags: [Pedidos]
      operationId: getOrder
      summary: Leer un pedido
      responses:
        "200":
          description: Pedido
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
  /orders/{orderId}/cartons:
    parameters:
      - $ref: "#/components/parameters/orderId"
    post:
      tags: [Pedidos]
      operationId: openCarton
      summary: Abrir un cartón
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CodeWrite"
      responses:
        "201":
          description: Cartón abierto
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Carton"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /orders/{orderId}/cartons/{cartonId}/picks:
    parameters:
      - $ref: "#/components/parameters/orderId"
      - $ref: "#/components/parameters/cartonId"
    post:
      tags: [Pedidos]
      operationId: pickIntoCarton
      summary: Retirar cantidad de un bin hacia el cartón
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PickWrite"
      responses:
        "201":
          description: Pick registrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Pick"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /orders/{orderId}/pack:
    parameters:
      - $ref: "#/components/parameters/orderId"
    post:
      tags: [Pedidos]
      operationId: packOrder
      summary: Embalar el pedido cuando cada línea está pickeada
      responses:
        "200":
          description: Pedido embalado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
components:
  parameters:
...
    productId:
      name: productId
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
...
    orderId:
      name: orderId
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
    cartonId:
      name: cartonId
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
  responses:
    BadRequest:
      description: El cuerpo o el id no cumplen el contrato
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: No existe ese recurso
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Conflict:
      description: El estado actual no permite la operación
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    Product:
      type: object
      required: [id, sku, name, categoryId]
      properties:
        id:
          type: integer
          minimum: 1
        sku:
          type: string
        name:
          type: string
        categoryId:
          type: integer
          minimum: 1
    ProductWrite:
      type: object
      required: [sku, name, categoryId]
      properties:
        sku:
          type: string
          minLength: 1
        name:
          type: string
          minLength: 1
        categoryId:
          type: integer
          minimum: 1
    ProductPatch:
      type: object
      minProperties: 1
      properties:
        sku:
          type: string
          minLength: 1
        name:
          type: string
          minLength: 1
        categoryId:
          type: integer
          minimum: 1
    ProductBalance:
      type: object
      required: [binId, quantity, binCode, shelfId, shelfCode, rowId, rowCode, aisleId, aisleCode, warehouseId, warehouseCode]
      properties:
        binId:
          type: integer
        quantity:
          type: integer
          minimum: 0
        binCode:
          type: string
        shelfId:
          type: integer
        shelfCode:
          type: string
        rowId:
          type: integer
        rowCode:
          type: string
        aisleId:
          type: integer
        aisleCode:
          type: string
        warehouseId:
          type: integer
        warehouseCode:
          type: string
    ProductStock:
      type: object
      required: [productId, total, balances]
      properties:
        productId:
          type: integer
        total:
          type: integer
          minimum: 0
        balances:
          type: array
          items:
            $ref: "#/components/schemas/ProductBalance"
...
    CodeWrite:
      type: object
      required: [code]
      properties:
        code:
          type: string
          minLength: 1
...
    OrderLineWrite:
      type: object
      required: [productId, quantity]
      properties:
        productId:
          type: integer
          minimum: 1
        quantity:
          type: integer
          minimum: 1
    OrderWrite:
      type: object
      required: [lines]
      properties:
        lines:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/OrderLineWrite"
    OrderLine:
      type: object
      required: [productId, quantity, picked]
      properties:
        productId:
          type: integer
        quantity:
          type: integer
          minimum: 1
        picked:
          type: integer
          minimum: 0
    Pick:
      type: object
      required: [id, productId, binId, quantity]
      properties:
        id:
          type: integer
        productId:
          type: integer
        binId:
          type: integer
        quantity:
          type: integer
          minimum: 1
    PickWrite:
      type: object
      required: [productId, binId, quantity]
      properties:
        productId:
          type: integer
          minimum: 1
        binId:
          type: integer
          minimum: 1
        quantity:
          type: integer
          minimum: 1
    Carton:
      type: object
      required: [id, code, picks]
      properties:
        id:
          type: integer
        code:
          type: string
        picks:
          type: array
          items:
            $ref: "#/components/schemas/Pick"
    Order:
      type: object
      required: [id, status, createdAt, lines, cartons]
      properties:
        id:
          type: integer
        status:
          type: string
          enum: [open, packed]
        createdAt:
          type: string
          format: date-time
        lines:
          type: array
          items:
            $ref: "#/components/schemas/OrderLine"
        cartons:
          type: array
          items:
            $ref: "#/components/schemas/Carton"